adds float support by more buil-in aggregation functions

+++ src/main/java/io/github/vmzakharov/ecdataframe/dataframe/AggregateFunction.java
@@ -231,7 +231,7 @@ public abstract class AggregateFunction
         throw this.unsupportedAccumulatorException(ValueType.DOUBLE.toString());
     }
 
-    protected int floatAccumulator(float currentAggregate, float newValue)
+    protected float floatAccumulator(float currentAggregate, float newValue)
     {
         throw this.unsupportedAccumulatorException(ValueType.FLOAT.toString());
     }
